fix licence in all files
Summary:
Added licence to the new files which had been added since the renaming when I added the header to all the original files.

Shortened the licence header to one line (as it is in detectron2 and other codebases).

This diff should fix the license automated check on the repo (combined with D19343514 which skips the .obj and .mtl files and the docs).

Reviewed By: gkioxari

Differential Revision: D19343568

fbshipit-source-id: 88ae842702a8e21617402e93ba2e4ae6b528f7ad