CircleCI testing
Summary:
Fix risk of picking up old torchvision. We need the docs to be present to run the tests due to reference to the cow mesh.

This makes python 3.8 work (note that there's no 'm' in the abi tag because pymalloc is compatible now). This is a bit of a cheat because torchvision is not released for python 3.8 yet.

Reviewed By: nikhilaravi

Differential Revision: D19469873

fbshipit-source-id: 98a7e4ab3adc35d7978239fa08804e7a710e4f0e