pointcloud structure
Summary:
Revisions to Poincloud data structure with added normals

The biggest changes form the previous version include:
a) If the user provides tensor inputs, we make no assumption about padding. Padding is only for internal use for us to convert from list to padded
b) If features are not provided or if the poincloud is empty, all forms of features are None. This is so that we don't waste memory on holding dummy tensors.

Reviewed By: nikhilaravi

Differential Revision: D19791851

fbshipit-source-id: 7e182f7bb14395cb966531653f6dd6b328fd999c