Fix function call in Transform3d documentation (#233)
Summary:
The documentation of `Transform3d` highlights that the class handles points as well as normals. However, `transform_points` is applied to points and normals in the documentation instead of using `transform_normals` for normals, which I believe was intended.

This pull request fixes this typo in the documentation.
Pull Request resolved: https://github.com/facebookresearch/pytorch3d/pull/233

Reviewed By: bottler

Differential Revision: D22287642

Pulled By: nikhilaravi

fbshipit-source-id: 0bc8754097b2e17a34fa3071319d00b78c3bc803