Update LICENSE name to BSD 3-Clause License. (#310)
Summary:
The README for the project refers to the BSD 3-Clause License, but the License file has only BSD License as the name. Updated the License name to BSD 3-Clause License for correct reference.

License files are an important part of source code and should not have errors. This helps developers to correctly reference the license when integrating the source code within their applications.

Pull Request resolved: https://github.com/facebookresearch/pytorch3d/pull/310

Reviewed By: theschnitz

Differential Revision: D24220599

Pulled By: nikhilaravi

fbshipit-source-id: 657bde0abe946e95864fc048eb930a91ddf08b99